pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "1.0.0"
description = ""
authors = ["ShcherbinaDmitry <shcherbinadmitry@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.6"
django-bootstrap5 = "^23.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
django-filter = "^23.5"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
coverage = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
